+++ wazo_auth/plugins/http/tokens/schemas.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2025 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2026 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from marshmallow import validates_schema
@@ -72,9 +72,15 @@ class TokenRequestSchema(BaseSchema):
 
 
 class RefreshTokenListSchema(BaseListSchema):
-    sort_columns = ['created_at', 'client_id', 'mobile']
+    sort_columns = ['created_at', 'client_id', 'mobile', 'last_used_at']
     default_sort_column = 'created_at'
-    searchable_columns = ['created_at', 'client_id', 'mobile']
+    searchable_columns = [
+        'created_at',
+        'client_id',
+        'mobile',
+        'user_uuid',
+        'user_agent',
+    ]
 
 
 class RefreshTokenSchema(BaseSchema):
@@ -83,6 +89,8 @@ class RefreshTokenSchema(BaseSchema):
     mobile = fields.Boolean()
     user_uuid = fields.String()
     tenant_uuid = fields.String()
+    user_agent = fields.String()
+    last_used_at = fields.DateTime()
     metadata = fields.Dict()